requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
pika==1.3.2
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.2.1
psycopg2==2.9.10
pyaes==1.6.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pyrogram==2.0.106
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
TgCrypto==1.2.5
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
